add loading overlay to details block
Refactored the loading overlay to move functions to shared-program.
Moved CSS to gmo-styles. Added to details block.
