+++ blocks/gmo-program-list/gmo-program-list.css
@@ -292,47 +292,3 @@ select {
     font: normal normal normal 12px/17px Adobe Clean;
     margin-left: 20px;
 }
-
-/* 'Loading' overlay */
-.loading-overlay {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(192, 192, 192, 0.701);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    z-index: 1000;
-}
-
-.spinner {
-    border: 4px solid rgba(255, 255, 255, 0.3);
-    border-top: 4px solid white;
-    border-radius: 50%;
-    width: 40px;
-    height: 40px;
-    animation: spin 1s linear infinite;
-}
-
-/* Keyframes for the spinner */
-@keyframes spin {
-    0% {
-        transform: rotate(0deg);
-    }
-    100% {
-        transform: rotate(360deg);
-    }
-}
-
-.loading-message {
-    color: white;
-    font-size: 1.2rem;
-    margin-left: 10px;
-}
-
-.loading-content {
-    display: flex;
-    align-items: center;
-}
